[SDSAET-38] Move StatusCodeError models to models-api module
