+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/statuscodes/StatusCodesComparator.java
@@ -21,9 +21,10 @@ import com.cognifide.aet.job.api.comparator.ComparatorProperties;
 import com.cognifide.aet.job.api.datafilter.DataFilterJob;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
-import com.cognifide.aet.job.common.collectors.statuscodes.StatusCode;
 import com.cognifide.aet.job.common.collectors.statuscodes.StatusCodesCollectorResult;
 import com.cognifide.aet.models.Excludable;
+import com.cognifide.aet.models.statuscodes.StatusCode;
+import com.cognifide.aet.models.statuscodes.StatusCodesComparatorResult;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Lists;
