+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/statuscodes/StatusCodesFilter.java
@@ -15,7 +15,7 @@
  */
 package com.cognifide.aet.job.common.comparators.statuscodes;
 
-import com.cognifide.aet.job.common.collectors.statuscodes.StatusCode;
+import com.cognifide.aet.models.statuscodes.StatusCode;
 import com.google.common.base.Predicate;
 import com.google.common.collect.FluentIterable;
 import java.util.List;
