+++ core/jobs/src/main/java/com/cognifide/aet/job/common/datafilters/statuscodesfilter/StatusCodesFilter.java
@@ -18,8 +18,8 @@ package com.cognifide.aet.job.common.datafilters.statuscodesfilter;
 import com.cognifide.aet.job.api.datafilter.AbstractDataModifierJob;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
-import com.cognifide.aet.job.common.collectors.statuscodes.StatusCode;
 import com.cognifide.aet.job.common.collectors.statuscodes.StatusCodesCollectorResult;
+import com.cognifide.aet.models.statuscodes.StatusCode;
 import java.util.Map;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
