+++ core/jobs/src/test/java/com/cognifide/aet/job/common/comparators/statuscodes/StatusCodesFilterTest.java
@@ -20,7 +20,7 @@ import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 
-import com.cognifide.aet.job.common.collectors.statuscodes.StatusCode;
+import com.cognifide.aet.models.statuscodes.StatusCode;
 import com.google.common.collect.Lists;
 import com.googlecode.zohhak.api.TestWith;
 import com.googlecode.zohhak.api.runners.ZohhakRunner;
